+++ src/main/java/com/hitwh/vblog/service/impl/ThumbServiceImpl.java
@@ -1,7 +1,9 @@
 package com.hitwh.vblog.service.impl;
 
+import com.hitwh.vblog.bean.ArticleDo;
 import com.hitwh.vblog.bean.ArticleDynamicDo;
 import com.hitwh.vblog.bean.ThumbRecordDo;
+import com.hitwh.vblog.mapper.ArticleDoMapper;
 import com.hitwh.vblog.mapper.ArticleDynamicDoMapper;
 import com.hitwh.vblog.mapper.ThumbRecordDoMapper;
 import com.hitwh.vblog.model.ThumbModel;
@@ -28,6 +30,8 @@ public class ThumbServiceImpl implements ThumbService {
     ValidatorImpl validator;
     @Autowired
     ArticleDynamicDoMapper articleDynamicDoMapper;
+    @Autowired
+    ArticleDoMapper articleDoMapper;
 
     @Override
     public void insertThumbRecord(ThumbModel thumbModel) throws BusinessException {
@@ -41,22 +45,49 @@ public class ThumbServiceImpl implements ThumbService {
         ThumbRecordDo thumbRecordDo = new ThumbRecordDo();
         thumbRecordDo.setArticleId(thumbModel.getArticleId());
         thumbRecordDo.setThumberId(thumbModel.getUserId());
+
+        ArticleDo articleDo = null;
+        try {
+            articleDo = articleDoMapper.selectByPrimaryKey(thumbModel.getArticleId());
+        }catch (Exception e){
+            throw new BusinessException(EnumError.ARTICLE_NOT_EXIST);
+        }
+
+        if(articleDo == null)
+            throw new BusinessException(EnumError.ARTICLE_NOT_EXIST);
+
         thumbRecordDo.setThumbTime(new Date(System.currentTimeMillis()));
 
-        Integer column = thumbRecordDoMapper.insertSelective(thumbRecordDo);
+        int num = thumbRecordDoMapper.selectIfThumb(thumbRecordDo);
+        if(num != 0 )
+            throw new BusinessException(EnumError.THUMB_EXIST);
+
+        Integer column = 0;
+        try {
+            column = thumbRecordDoMapper.insertSelective(thumbRecordDo);
+        }catch (Exception e){
+            System.out.println("插入点赞失败");
+            throw new BusinessException(EnumError.THUMB_INSERT_ERROR);
+        }
 
         if(column == null || column == 0)
-            throw new BusinessException(EnumError.INTERNAL_SERVER_ERROR);
+            throw new BusinessException(EnumError.THUMB_INSERT_ERROR);
 
         ArticleDynamicDo articleDynamicDo = new ArticleDynamicDo();
         articleDynamicDo.setArticleId(thumbModel.getArticleId());
         articleDynamicDo.setThumbNum(1);
 
         // 点赞数加一
-        Integer addColum = articleDynamicDoMapper.addArticleDynamic(articleDynamicDo);
+        Integer addColum = 0;
+        try {
+            addColum = articleDynamicDoMapper.addArticleDynamic(articleDynamicDo);
+        }catch (Exception e){
+            System.out.println("点赞数加一失败");
+            throw new BusinessException(EnumError.THUMB_INSERT_ERROR);
+        }
 
         if(addColum == null || addColum == 0)
-            throw new BusinessException(EnumError.INTERNAL_SERVER_ERROR);
+            throw new BusinessException(EnumError.THUMB_INSERT_ERROR);
     }
 
     @Override
@@ -64,9 +95,16 @@ public class ThumbServiceImpl implements ThumbService {
         if(articleId == null || articleId == 0)
             throw new BusinessException(EnumError.PARAMETER_VALIDATION_ERROR, "传入参数错误");
 
-        ArticleDynamicDo articleDynamicDo = articleDynamicDoMapper.selectByPrimaryKey(articleId);
+        ArticleDynamicDo articleDynamicDo = null;
+        try {
+            articleDynamicDo = articleDynamicDoMapper.selectByPrimaryKey(articleId);
+        }catch (Exception e){
+            throw new BusinessException(EnumError.ARTICLE_NOT_EXIST);
+        }
+
         if(articleDynamicDo == null || articleDynamicDo.getThumbNum() == null )
-            throw  new BusinessException(EnumError.INTERNAL_SERVER_ERROR);
+            throw new BusinessException(EnumError.ARTICLE_NOT_EXIST);
+
         return articleDynamicDo.getThumbNum();
     }
 
@@ -83,21 +121,44 @@ public class ThumbServiceImpl implements ThumbService {
         ThumbRecordDo thumbRecordDo = new ThumbRecordDo();
         thumbRecordDo.setArticleId(thumbModel.getArticleId());
         thumbRecordDo.setThumberId(thumbModel.getUserId());
+        ArticleDo articleDo = articleDoMapper.selectByPrimaryKey(thumbModel.getArticleId());
+        if(articleDo == null)
+            throw new BusinessException(EnumError.ARTICLE_NOT_EXIST);
 
-        Integer column = thumbRecordDoMapper.delete(thumbRecordDo);
+        int num = thumbRecordDoMapper.selectIfThumb(thumbRecordDo);
+        if(num == 0 )
+            throw new BusinessException(EnumError.THUMB_NOT_EXIST);
+
+        Integer column = 0;
+        try {
+            column = thumbRecordDoMapper.delete(thumbRecordDo);
+        }catch (Exception e){
+            System.out.println("点赞数加一失败");
+            throw new BusinessException(EnumError.THUMB_DELETE_ERROR);
+        }
+
 
         if(column == null || column == 0)
-            throw new BusinessException(EnumError.INTERNAL_SERVER_ERROR);
+            throw new BusinessException(EnumError.THUMB_DELETE_ERROR);
 
         ArticleDynamicDo articleDynamicDo = new ArticleDynamicDo();
         articleDynamicDo.setArticleId(thumbModel.getArticleId());
         articleDynamicDo.setThumbNum(1);
 
         // 点赞数减一
-        Integer addColum = articleDynamicDoMapper.subtractArticleDynamic(articleDynamicDo);
+        Integer addColum = 0;
+        try{
+            addColum = articleDynamicDoMapper.subtractArticleDynamic(articleDynamicDo);
+        }catch (Exception e){
+            System.out.println("点赞数减一失败");
+            throw new BusinessException(EnumError.THUMB_DELETE_ERROR);
+        }
 
-        if(addColum == null || addColum == 0)
-            throw new BusinessException(EnumError.INTERNAL_SERVER_ERROR);
+
+        if(addColum == null || addColum == 0){
+            System.out.println("点赞数减一失败");
+            throw new BusinessException(EnumError.THUMB_DELETE_ERROR);
+        }
 
     }
 
@@ -114,9 +175,9 @@ public class ThumbServiceImpl implements ThumbService {
         thumbRecordDo.setArticleId(thumbModel.getArticleId());
         thumbRecordDo.setThumberId(thumbModel.getUserId());
 
-        ThumbRecordDo recordDo = thumbRecordDoMapper.selectIfThumb(thumbRecordDo);
+        int num = thumbRecordDoMapper.selectIfThumb(thumbRecordDo);
 
-        if(recordDo == null)
+        if(num == 0)
             return false;
         else
             return true;
